fix: corrige erros de inicialização e seed dos cursos iniciais
- JwtService: usa bytes raw da secret em vez de Base64.decode (fix 500 no register)
- RedisConfig: remove .transactionAware() que causava conflito com @Transactional
- CourseRating: campo rating como Short/SMALLINT para alinhar com schema
- RatingService: cast para (short) ao setar rating
- Course: columnDefinition NUMERIC(3,2) em ratingAvg para alinhar com schema
- application.yml: credenciais DB via env vars, cookie-secure=false por padrão
- application-dev.yml: cache.type=simple para rodar sem Redis obrigatório
- V9: migration que renomeia payload→data e adiciona title/message em notifications
- V10: seed do curso Tribo do Investidor (13 módulos, 56 aulas)
- V11: seed do curso Organização Financeira e Negociação de Dívidas (3 módulos, 8 aulas)
- pom.xml: restaurado após corrupção acidental

+++ src/main/java/com/tribo/modules/course/service/RatingService.java
@@ -42,7 +42,7 @@ public class RatingService {
                         .userId(userId)
                         .build());
 
-        rating.setRating(stars);
+        rating.setRating((short) stars);
         rating.setReview(review);
         rating.setUpdatedAt(OffsetDateTime.now());
         ratingRepository.save(rating);
